Make rc channel numbers and calibration params use int instead of float As parameter subsystems supports int now, and int is much more natural here.

+++ flix/parameters.ino
@@ -6,9 +6,9 @@
 #include <Preferences.h>
 #include "util.h"
 
-extern float channelZero[16];
-extern float channelMax[16];
-extern float rollChannel, pitchChannel, throttleChannel, yawChannel, armedChannel, modeChannel;
+extern int channelZero[16];
+extern int channelMax[16];
+extern int rollChannel, pitchChannel, throttleChannel, yawChannel, armedChannel, modeChannel;
 extern int wifiMode, udpLocalPort, udpRemotePort;
 extern float rcLossTimeout, descendTime;
